Reworked OrderItem so that data is pulled automatically via gRPC, rather than trusting data from JSON (Clothe Info).

+++ Clothy.Services/Clothy.OrderService/Clothy.OrderService.DAL/Repositories/OrderRepository.cs
@@ -91,7 +91,7 @@ namespace Clothy.OrderService.DAL.Repositories
                 UserLastName = r.userlastname,
                 CreatedAt = r.createdat,
                 UpdatedAt = r.updatedat,
-                TotalAmount = (double)r.totalamount,
+                TotalAmount = (decimal)r.totalamount,
                 Status = new OrderStatus
                 {
                     Id = r.statusid,
