+++ Clothy.Services/Clothy.OrderService/Clothy.OrderService.Domain/Entities/AdditionalEntities/OrderWithDetailsData.cs
@@ -12,7 +12,7 @@ namespace Clothy.OrderService.Domain.Entities.AdditionalEntities
         public OrderStatus? Status { get; set; } 
         public string? UserFirstName { get; set; }
         public string? UserLastName { get; set; }
-        public double TotalAmount { get; set; }
+        public decimal TotalAmount { get; set; }
         public List<OrderItemData> Items { get; set; } = new List<OrderItemData>();
         public DeliveryDetailData? DeliveryDetail { get; set; }
         public DateTime CreatedAt { get; set; }
